# docs/specs/auth-service-openapi.yml
openapi: 3.0.3
info:
  title: Shard Legends Auth Service API
  description: |
    Микросервис авторизации и аутентификации для игры Shard Legends: Clan Wars.
    
    Обеспечивает авторизацию пользователей через Telegram Web App, 
    генерацию JWT токенов и управление сессиями.
    
    **Основные возможности:**
    - Валидация данных Telegram Web App через HMAC-SHA256
    - Автоматическая регистрация новых пользователей
    - Генерация JWT токенов с RSA подписью (24 часа)
    - Управление активными/отозванными токенами в Redis
    - Rate limiting для защиты от злоупотреблений
  version: 1.0.0
  contact:
    name: Shard Legends Development Team
  license:
    name: MIT
servers:
  - url: https://dev.slcw.dimlight.online/api
    description: Development server
  - url: https://stage.slcw.dimlight.online/api
    description: Staging server  
  - url: https://slcw.dimlight.online/api
    description: Production server
...
paths:
...
  /admin/tokens/stats:
    get:
      tags:
        - Admin
      summary: Статистика активных токенов
      description: |
        Получение статистики по активным токенам и информации о последней очистке.
        
        **Использование:** Мониторинг состояния токенов и производительности системы.
      operationId: getTokenStats
      responses:
        '200':
          description: Статистика токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenStatsResponse'
              example:
                active_tokens: 1247
                unique_users: 892
                cleanup_stats:
                  last_cleanup: "2024-12-21T15:30:00Z"
                  expired_tokens_cleaned: 156
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/tokens/user/{userId}:
    get:
      tags:
        - Admin
      summary: Получение активных токенов пользователя
      description: Получение списка всех активных токенов для указанного пользователя
      operationId: getUserTokens
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список активных токенов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTokensResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Admin
      summary: Отзыв всех токенов пользователя
      description: Отзывает все активные токены указанного пользователя
      operationId: revokeUserTokens
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Токены успешно отозваны
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeTokensResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/tokens/{jti}:
    delete:
      tags:
        - Admin
      summary: Отзыв конкретного токена
      description: Отзывает токен по его JTI (JWT ID)
      operationId: revokeToken
      parameters:
        - name: jti
          in: path
          required: true
          description: JWT ID токена
          schema:
            type: string
      responses:
        '200':
          description: Токен успешно отозван
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeTokenResponse'
        '404':
          description: Токен не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          description: Статус операции (всегда false для ошибок)
          example: false
        error:
          type: string
          description: Код ошибки для программной обработки
          enum:
            - missing_init_data
            - invalid_init_data_format
            - expired_auth_date
            - invalid_telegram_signature
            - invalid_user_data
            - rate_limit_exceeded
            - internal_server_error
          example: "invalid_telegram_signature"
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: "HMAC-SHA256 signature verification failed"
        details:
          type: object
          description: Дополнительные детали ошибки (опционально)
          additionalProperties: true
...
    TokenStatsResponse:
      type: object
      required:
        - active_tokens
        - unique_users
      properties:
        active_tokens:
          type: integer
          description: Количество активных токенов в системе
          example: 1247
        unique_users:
          type: integer
          description: Количество уникальных пользователей с активными токенами
          example: 892
        cleanup_stats:
          type: object
          description: Статистика последней очистки токенов
          properties:
            last_cleanup:
              type: string
              format: date-time
              description: Время последней автоматической очистки
              example: "2024-12-21T15:30:00Z"
            expired_tokens_cleaned:
              type: integer
              description: Количество удаленных просроченных токенов при последней очистке
              example: 156
    UserTokensResponse:
      type: object
      required:
        - user_id
        - active_tokens
      properties:
        user_id:
          type: string
          format: uuid
          description: UUID пользователя
          example: "555aaa-bbbb-cccc-dddd-eeeeeeeeeeee"
        active_tokens:
          type: array
          description: Список активных токенов пользователя
          items:
            type: object
            properties:
              jti:
                type: string
                description: JWT ID токена
                example: "12345-67890-abcdef"
              issued_at:
                type: string
                format: date-time
                description: Время выдачи токена
                example: "2024-12-21T10:30:00Z"
              expires_at:
                type: string
                format: date-time
                description: Время истечения токена
                example: "2024-12-22T10:30:00Z"
    RevokeTokensResponse:
      type: object
      required:
        - success
        - revoked_tokens
        - user_id
      properties:
        success:
          type: boolean
          description: Статус операции
          example: true
        revoked_tokens:
          type: integer
          description: Количество отозванных токенов
          example: 3
        user_id:
          type: string
          format: uuid
          description: UUID пользователя
          example: "555aaa-bbbb-cccc-dddd-eeeeeeeeeeee"
    RevokeTokenResponse:
      type: object
      required:
        - success
        - revoked_token
      properties:
        success:
          type: boolean
          description: Статус операции
          example: true
        revoked_token:
          type: string
          description: JTI отозванного токена
          example: "12345-67890-abcdef"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен, полученный от эндпоинта /auth.
        
        Формат заголовка: `Authorization: Bearer <token>`
        
        **Примечание:** Данная схема безопасности используется другими сервисами
        для валидации токенов, выданных auth-service.
security: []
...